Viridi-batteryIssues-Reporting/scripts/requirement.txt
pyspark==2.3.4
boto3==1.15.18
asn1crypto==1.5.1
azure-common==1.1.28
azure-core==1.26.0
azure-storage-blob==12.14.0
beautifulsoup4==4.9.3
boto==2.49.0
botocore==1.18.18
certifi==2020.12.5
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.24
fonttools==4.37.4
idna==2.10
ipaddress==1.0.23
isodate==0.6.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.4.4
lxml==4.6.2
matplotlib==3.5.3
msrest==0.7.1
mysqlclient==1.4.2
nltk==3.5
nose==1.3.4
numpy==1.21.6
oauthlib==3.2.1
oscrypto==1.3.0
packaging==21.3
pandas==1.3.3
Pillow==9.2.0
py-dateutil==2.2
pyarrow==0.17.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2020.5
PyYAML==5.4.1
regex==2021.3.17
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.3.0
semantic-version==2.10.0
setuptools-rust==1.5.2
six==1.13.0
snowflake-connector-python==2.3.9
tqdm==4.59.0
typing_extensions==4.4.0
urllib3==1.21.1
windmill==1.6
sql-metadata